# Eliminate unnecessary Bitmap allocation in (*Bitmap).lazyOR