Fixes #138

Benchmark on the "real" dataset show a significant improvement:

```
benchmark                                            old ns/op     new ns/op     delta
BenchmarkRealDataFastOr/census-income_srt-4          841099        736937        -12.38%
BenchmarkRealDataFastOr/census-income-4              1248805       1157445       -7.32%
BenchmarkRealDataFastOr/census1881_srt-4             1992651       1295663       -34.98%
BenchmarkRealDataFastOr/census1881-4                 2797780       2096418       -25.07%
BenchmarkRealDataFastOr/dimension_003-4              50486750      2174251       -95.69%
BenchmarkRealDataFastOr/dimension_008-4              17056898      904640        -94.70%
BenchmarkRealDataFastOr/dimension_033-4              1246254       794440        -36.25%
BenchmarkRealDataFastOr/uscensus2000-4               5387897       2702060       -49.85%
BenchmarkRealDataFastOr/weather_sept_85_srt-4        713331        505671        -29.11%
BenchmarkRealDataFastOr/weather_sept_85-4            5248820       4952945       -5.64%
BenchmarkRealDataFastOr/wikileaks-noquotes_srt-4     567737        276666        -51.27%
BenchmarkRealDataFastOr/wikileaks-noquotes-4         906403        613208        -32.35%
```