@lemire 2018-02-06
Looks safe enough, and sensible.